requirements.txt
fastapi
uvicorn
requests
python-dotenv
youtube-transcript-api
pydantic